+++ beam/beam-visat-rcp/src/main/java/org/esa/beam/visat/actions/AttachPixelGeoCodingAction.java
@@ -102,8 +102,8 @@ public class AttachPixelGeoCodingAction extends ExecCommand {
         if (freeMegas < requiredMegas) {
             // TODO - make this a common dialog, e.g. for RGB image creation etc
             final String message = MessageFormat.format("This operation requires to load at least {0} M\n" +
-                    "of additional data into memory.\n\n" +
-                    "Do you really want to continue?",
+                                                                "of additional data into memory.\n\n" +
+                                                                "Do you really want to continue?",
                                                         requiredMegas);   /*I18N*/
             final int answer = visatApp.showQuestionDialog(ATTACH_TITLE,
                                                            message, null);
@@ -168,7 +168,7 @@ public class AttachPixelGeoCodingAction extends ExecCommand {
         private JSpinner _radiusSpinner;
         private final Integer _defaultRadius = new Integer(6);
         private final Integer _minRadius = new Integer(0);
-        private final Integer _maxRadius = new Integer(10);
+        private final Integer _maxRadius = new Integer(99);
         private final Integer _bigRadiusStep = new Integer(0);
         private final Integer _smallRadiusStep = new Integer(1);
